added new launch config to complex and single paw tiling. Used old version of paw tiling for outermost loop parallelization

+++ resources/examples/tiling/MatrixMultiplication/NonTiled.c
@@ -22,7 +22,7 @@ int main(int argc, char const *argv[])
     }
 
 
-    int a[n][n], b[n][m], d[n][m];
+    float a[n][n], b[n][m], d[n][m];
 
 
     int i, j, k;
