+++ resources/examples/tiling/MatrixMultiplication/ParallelNonTiled.c
@@ -23,7 +23,7 @@ int main(int argc, char const *argv[])
     }
 
 
-    int a[n][n], b[n][m], d[n][m];
+    float a[n][n], b[n][m], d[n][m];
 
 
     int i, j, k;
