+++ resources/examples/tiling/MatrixVectorMultiplication/NonTiled.c
@@ -16,7 +16,7 @@ int main(int argc, char const *argv[])
         n = argv[0];
     }
 
-    int a[n*n], b[n], c[n];
+    float a[n*n], b[n], c[n];
 
     int i, j;
